relation of more than one class
